Move some StringUtil methods to CollectionUtil
